Refactor decoration innerZ handling and update documentation
- Changed the innerZ parameter in decoration methods to be signed, allowing for negative values that draw decorations behind the parent entity.
- Updated related constants in ConfigDefaults.js to define the range for innerZ.
- Adjusted the documentation to reflect the new behavior of innerZ and its impact on decoration rendering.

+++ src/components/DecorationComponent.js
@@ -44,7 +44,7 @@ export class DecorationComponent extends Component {
     localX: Float32Array,
     localY: Float32Array,
     inheritParentRotation: Uint8Array, // 1 = add parent Transform.rotation to baseRotation for display
-    innerZ: Uint16Array, // 0..1023 sub-layer for Y-sort (composite = parentY*1024 + innerZ)
+    innerZ: Int8Array, // signed; clamp to DECORATION_INNER_Z_MIN..MAX; composite = worldY*SCALE + innerZ
   };
 
   // Static pool tracking (set during initialization)
